# Use stress divergence and diffusive flux divergence from Turbulence Closures