--- a/src/closures/closure_operators.jl
+++ b/src/closures/closure_operators.jl
@@ -3,10 +3,12 @@
 #
 
 ∂x_caa(i, j, k, grid, u::AbstractArray) = δx_f2c(grid, u, i, j, k) / grid.Δx
-∂y_aca(i, j, k, grid, v::AbstractArray) = δy_f2c(grid, v, i, j, k) / grid.Δy
-∂z_aac(i, j, k, grid, w::AbstractArray) = δz_f2c(grid, w, i, j, k) / grid.Δz
 ∂x_faa(i, j, k, grid, ϕ::AbstractArray) = δx_c2f(grid, ϕ, i, j, k) / grid.Δx
+
+∂y_aca(i, j, k, grid, v::AbstractArray) = δy_f2c(grid, v, i, j, k) / grid.Δy
 ∂y_afa(i, j, k, grid, ϕ::AbstractArray) = δy_c2f(grid, ϕ, i, j, k) / grid.Δy
+
+∂z_aac(i, j, k, grid, w::AbstractArray) = δz_f2c(grid, w, i, j, k) / grid.Δz
 ∂z_aaf(i, j, k, grid, ϕ::AbstractArray) = δz_c2f(grid, ϕ, i, j, k) / grid.Δz
 
 #
@@ -16,22 +18,22 @@
 #
 # face   cell   face   cell   face
 #
-#         i-1            i 
+#         i-1            i
 #          ↓             ↓
 #   |      ×      |      ×      |
 #   ↑             ↑             ↑
 #  i-1            i            i+1
 #
-# As a result the interpolation of a quantity ϕᵢ from a cell i to face i 
-# (this operation is denoted "▶x_faa" in the code below) is 
+# As a result the interpolation of a quantity ϕᵢ from a cell i to face i
+# (this operation is denoted "▶x_faa" in the code below) is
 #
 # (▶ϕ)ᵢ = (ϕᵢ + ϕᵢ₋₁) / 2 .
 #
 # Conversely, the interpolation of a quantity uᵢ from face i to cell i is given by
 #
 # (▶u)ᵢ = (uᵢ₊₁ + uᵢ) / 2.
-# 
-# Derivative operators are defined similarly. Using the symbol "∂" to denote 
+#
+# Derivative operators are defined similarly. Using the symbol "∂" to denote
 # differentiation, we have
 #
 # (∂ϕ)ᵢ = (ϕᵢ - ϕᵢ₋₁) / Δ
@@ -42,7 +44,7 @@
 #
 # for the differentation of a FaceField ending on cell i, where Δ is the spacing
 # between cell points or face points, appropriately.
-# 
+#
 
 #
 # Differential operators
@@ -114,7 +116,7 @@ Differentiate the function or callable object
 located at `aac` in `z`, across `aaf`.
 """
 function ∂z_aaf(i, j, k, grid::Grid{T}, F::Function, args...) where T
-    if k < 2
+    if k == 1
         return -zero(T)
     else
         return (F(i, j, k-1, grid, args...) - F(i, j, k, grid, args...)) / grid.Δz
@@ -130,9 +132,9 @@ Differentiate the function or callable object
 
 located at `aaf` in `z`, across `aac`.
 """
-function ∂z_aac(i, j, k, grid::Grid{T}, F::Function, args...) where T
+function ∂z_aac(i, j, k, grid, F::Function, args...)
     if k == grid.Nz
-        return -zero(T)
+        return F(i, j, k, grid, args...) / grid.Δz
     else
         return (F(i, j, k, grid, args...) - F(i, j, k+1, grid, args...)) / grid.Δz
     end
@@ -142,14 +144,14 @@ end
 # Double differentiation
 #
 
-∂x²_caa(i, j, k, grid, u) = ∂x_faa(i, j, k, grid, ∂x_caa, u)
-∂x²_faa(i, j, k, grid, ϕ) = ∂x_caa(i, j, k, grid, ∂x_faa, ϕ)
+∂x²_caa(i, j, k, grid, ϕ) = ∂x_caa(i, j, k, grid, ∂x_faa, ϕ)
+∂x²_faa(i, j, k, grid, u) = ∂x_faa(i, j, k, grid, ∂x_caa, u)
 
-∂y²_aca(i, j, k, grid, v) = ∂y_afa(i, j, k, grid, ∂y_aca, v)
-∂y²_afa(i, j, k, grid, ϕ) = ∂y_aca(i, j, k, grid, ∂y_afa, ϕ)
+∂y²_aca(i, j, k, grid, ϕ) = ∂y_aca(i, j, k, grid, ∂y_afa, ϕ)
+∂y²_afa(i, j, k, grid, v) = ∂y_afa(i, j, k, grid, ∂y_aca, v)
 
-∂z²_aac(i, j, k, grid, w) = ∂z_aaf(i, j, k, grid, ∂z_aac, w)
-∂z²_aaf(i, j, k, grid, ϕ) = ∂z_aac(i, j, k, grid, ∂z_aaf, ϕ)
+∂z²_aac(i, j, k, grid, ϕ) = ∂z_aac(i, j, k, grid, ∂z_aaf, ϕ)
+∂z²_aaf(i, j, k, grid, w) = ∂z_aaf(i, j, k, grid, ∂z_aac, w)
 
 #
 # Interpolation operations for functions
@@ -389,146 +391,23 @@ in `y` and `z`, from `afc` to `acf`.
 """
 ▶yz_acf(i, j, k, grid, F, args...) = ▶z_aaf(i, j, k, grid, ▶y_aca, F, args...)
 
-"""
-    ∂x_caa(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_x ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `faa`, where `a` is either `f` or `c`.
-The quantity returned has the location `caa`.
-"""
-function ∂x_caa(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    i⁺¹ = incmod1(i, grid.Nx)
-
-    Σ⁺¹ = Σ(i⁺¹, j, k, grid, u, v, w)
-    Σ⁻¹ = Σ(i,   j, k, grid, u, v, w)
-
-    κ⁺¹ = κ(i⁺¹, j, k, grid, closure, eos, g, u, v, w, T, S)
-    κ⁻¹ = κ(i,   j, k, grid, closure, eos, g, u, v, w, T, S)
-
-    return (κ⁺¹ * Σ⁺¹ - κ⁻¹ * Σ⁻¹) / grid.Δx
-end
-
-"""
-    ∂x_faa(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_x ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `caa`, where `a` is either `f` or `c`.
-The quantity returned has the location `faa`.
-"""
-function ∂x_faa(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    i⁻¹ = decmod1(i, grid.Nx)
-
-    Σ⁺¹ = Σ(i,   j, k, grid, u, v, w)
-    Σ⁻¹ = Σ(i⁻¹, j, k, grid, u, v, w)
-
-    κ⁺¹ = κ(i,   j, k, grid, closure, eos, g, u, v, w, T, S)
-    κ⁻¹ = κ(i⁻¹, j, k, grid, closure, eos, g, u, v, w, T, S)
-
-    return (κ⁺¹ * Σ⁺¹ - κ⁻¹ * Σ⁻¹) / grid.Δx
-end
-
-"""
-    ∂y_aca(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_y ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `afa`, where `a` is either `f` or `c`.
-The quantity returned has the location `aca`.
-"""
-function ∂y_aca(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    j⁺¹ = incmod1(j, grid.Ny)
-
-    Σ⁺¹ = Σ(i, j⁺¹, k, grid, u, v, w)
-    Σ⁻¹ = Σ(i, j,   k, grid, u, v, w)
-
-    κ⁺¹ = κ(i, j⁺¹, k, grid, closure, eos, g, u, v, w, T, S)
-    κ⁻¹ = κ(i, j,   k, grid, closure, eos, g, u, v, w, T, S)
-
-    return (κ⁺¹ * Σ⁺¹ - κ⁻¹ * Σ⁻¹) / grid.Δy
-end
-
-"""
-    ∂y_afa(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_y ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `aca`, where `a` is either `f` or `c`.
-The quantity returned has the location `afa`.
-"""
-function ∂y_afa(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    j⁻¹ = decmod1(j, grid.Ny)
-
-    Σ⁺¹ = Σ(i, j,   k, grid, u, v, w)
-    Σ⁻¹ = Σ(i, j⁻¹, k, grid, u, v, w)
-
-    κ⁺¹ = κ(i, j,   k, grid, closure, eos, g, u, v, w, T, S)
-    κ⁻¹ = κ(i, j⁻¹, k, grid, closure, eos, g, u, v, w, T, S)
-
-    return (κ⁺¹ * Σ⁺¹ - κ⁻¹ * Σ⁻¹) / grid.Δy
-end
-
-"""
-    ∂z_aac(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_z ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `aaf`, where `a` is either `f` or `c`.
-The quantity returned has the location `aac`.
-"""
-function ∂z_aac(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    Σ⁻¹ = Σ(i, j, k, grid, u, v, w)
-    κ⁻¹ = κ(i, j, k, grid, closure, eos, g, u, v, w, T, S)
-
-    if k == grid.Nz
-        return κ⁻¹ * Σ⁻¹ / grid.Δz
-    else
-        Σ⁺¹ = Σ(i, j, k+1, grid, u, v, w)
-        κ⁺¹ = κ(i, j, k+1, grid, closure, eos, g, u, v, w, T, S)
-
-        # Reversed due to inverse convention for z-indexing
-        return (κ⁻¹ * Σ⁻¹ - κ⁺¹ * Σ⁺¹) / grid.Δz
-    end
-end
-
-"""
-    ∂z_aaf(κ, Σ, i, j, k, grid, closure, u, v, w, T, S)
-
-Compute ``\\partial_z ( \\kappa \\Sigma )``, where `κ` and `Σ`
-are diffusivity and strain functions that return quantities
-at the location `aac`, where `a` is either `f` or `c`.
-The quantity returned has the location `aaf`.
-"""
-function ∂z_aaf(κ::Function, Σ::Function, i, j, k, grid, closure, eos, g, u, v, w, T, S)
-    Σ⁺¹ = Σ(i, j, k, grid, u, v, w)
-    κ⁺¹ = κ(i, j, k, grid, closure, eos, g, u, v, w, T, S)
-
-    if k == 1
-        return - κ⁺¹ * Σ⁺¹ / grid.Δz
-    else
-        Σ⁻¹ = Σ(i, j, k-1, grid, u, v, w)
-        κ⁻¹ = κ(i, j, k-1, grid, closure, eos, g, u, v, w, T, S)
-
-        # Reversed due to inverse convention for z-indexing
-        return (κ⁻¹ * Σ⁻¹ - κ⁺¹ * Σ⁺¹) / grid.Δz
-    end
-end
+ν_Σᵢⱼ(i, j, k, grid, ν, Σᵢⱼ, closure, eos, g, u, v, w, T, S) =
+    ν(i, j, k, grid, closure, eos, g, u, v, w, T, S) * Σᵢⱼ(i, j, k, grid, u, v, w)
 
 # At fcc
-∂x_2ν_Σ₁₁(ijk...) = 2 * ∂x_faa(ν_ccc, Σ₁₁, ijk...)
-∂y_2ν_Σ₁₂(ijk...) = 2 * ∂y_aca(ν_ffc, Σ₁₂, ijk...)
-∂z_2ν_Σ₁₃(ijk...) = 2 * ∂z_aac(ν_fcf, Σ₁₃, ijk...)
+∂x_2ν_Σ₁₁(i, j, k, grid, args...) = 2 * ∂x_faa(i, j, k, grid, ν_Σᵢⱼ, ν_ccc, Σ₁₁, args...)
+∂y_2ν_Σ₁₂(i, j, k, grid, args...) = 2 * ∂y_aca(i, j, k, grid, ν_Σᵢⱼ, ν_ffc, Σ₁₂, args...)
+∂z_2ν_Σ₁₃(i, j, k, grid, args...) = 2 * ∂z_aac(i, j, k, grid, ν_Σᵢⱼ, ν_fcf, Σ₁₃, args...)
 
-# At cfc
-∂x_2ν_Σ₂₁(ijk...) = 2 * ∂x_caa(ν_ffc, Σ₂₁, ijk...)
-∂y_2ν_Σ₂₂(ijk...) = 2 * ∂y_afa(ν_ccc, Σ₂₂, ijk...)
-∂z_2ν_Σ₂₃(ijk...) = 2 * ∂z_aac(ν_cff, Σ₂₃, ijk...)
+# At cfcᵢⱼ
+∂x_2ν_Σ₂₁(i, j, k, grid, args...) = 2 * ∂x_caa(i, j, k, grid, ν_Σᵢⱼ, ν_ffc, Σ₂₁, args...)
+∂y_2ν_Σ₂₂(i, j, k, grid, args...) = 2 * ∂y_afa(i, j, k, grid, ν_Σᵢⱼ, ν_ccc, Σ₂₂, args...)
+∂z_2ν_Σ₂₃(i, j, k, grid, args...) = 2 * ∂z_aac(i, j, k, grid, ν_Σᵢⱼ, ν_cff, Σ₂₃, args...)
 
-# At ccf
-∂x_2ν_Σ₃₁(ijk...) = 2 * ∂x_caa(ν_fcf, Σ₃₁, ijk...)
-∂y_2ν_Σ₃₂(ijk...) = 2 * ∂y_aca(ν_cff, Σ₃₂, ijk...)
-∂z_2ν_Σ₃₃(ijk...) = 2 * ∂z_aaf(ν_ccc, Σ₃₃, ijk...)
+# At ccfᵢⱼ
+∂x_2ν_Σ₃₁(i, j, k, grid, args...) = 2 * ∂x_caa(i, j, k, grid, ν_Σᵢⱼ, ν_fcf, Σ₃₁, args...)
+∂y_2ν_Σ₃₂(i, j, k, grid, args...) = 2 * ∂y_aca(i, j, k, grid, ν_Σᵢⱼ, ν_cff, Σ₃₂, args...)
+∂z_2ν_Σ₃₃(i, j, k, grid, args...) = 2 * ∂z_aaf(i, j, k, grid, ν_Σᵢⱼ, ν_ccc, Σ₃₃, args...)
 
 """
     ∂ⱼ_2ν_Σ₁ⱼ(i, j, k, grid, closure, eos, g, u, v, w, T, S)
@@ -608,10 +487,14 @@ Return `κ ∂z ϕ`, where `κ` is a function that computes
 diffusivity at cell centers (location `ccc`), and `ϕ` is an array of scalar
 data located at cell centers.
 """
-function κ_∂z_ϕ(i, j, k, grid, ϕ, κ, closure, args...)
-    κ = ▶z_aaf(i, j, k, grid, κ, closure, args...)
-    ∂z_ϕ = ∂z_aaf(i, j, k, grid, ϕ)
-    return κ * ∂z_ϕ
+function κ_∂z_ϕ(i, j, k, grid::Grid{T}, ϕ, κ, closure, args...) where T
+    if k == 1
+        return -zero(T)
+    else
+        κ = ▶z_aaf(i, j, k, grid, κ, closure, args...)
+        ∂z_ϕ = ∂z_aaf(i, j, k, grid, ϕ)
+        return κ * ∂z_ϕ
+    end
 end
 
 """