--- a/src/closures/constant_diffusivity_closures.jl
+++ b/src/closures/constant_diffusivity_closures.jl
@@ -69,7 +69,6 @@ end
 ConstantAnisotropicDiffusivity(T; kwargs...) =
     typed_keyword_constructor(T, ConstantAnisotropicDiffusivity; kwargs...)
 
-#=
 @inline ∂ⱼ_2ν_Σ₁ⱼ(i, j, k, grid, closure::ConstantAnisotropicDiffusivity, eos, g, u, v, w, T, S) = (
       closure.νh/grid.Δx^2 * δx²_f2c2f(grid, u, i, j, k) 
     + closure.νh/grid.Δy^2 * δy²_f2e2f(grid, u, i, j, k) 
@@ -93,14 +92,8 @@ ConstantAnisotropicDiffusivity(T; kwargs...) =
     + closure.κh/g.Δy^2 * δy²_c2f2c(g, Q, i, j, k) 
     + closure.κv/g.Δz^2 * δz²_c2f2c(g, Q, i, j, k)
 )
-=#
 
-@inline ∇_κ_∇ϕ(i, j, k, grid, ϕ, closure::ConstantAnisotropicDiffusivity, args...) = (
-      ∂x_κ_∂x_ϕ(i, j, k, grid, ϕ, closure.κh, closure, args...)
-    + ∂y_κ_∂y_ϕ(i, j, k, grid, ϕ, closure.κh, closure, args...)
-    + ∂z_κ_∂z_ϕ(i, j, k, grid, ϕ, closure.κv, closure, args...)
-    )
-
+#=
 @inline ∂ⱼ_2ν_Σ₁ⱼ(i, j, k, grid, closure::ConstantAnisotropicDiffusivity, eos, g, u, v, w, T, S) = (
       closure.νh * ∂x²_faa(i, j, k, grid, u)
     + closure.νh * ∂y²_aca(i, j, k, grid, u)
@@ -118,6 +111,13 @@ ConstantAnisotropicDiffusivity(T; kwargs...) =
     + closure.νh * ∂y²_aca(i, j, k, grid, w)
     + closure.νv * ∂z²_aaf(i, j, k, grid, w)
     )
+
+@inline ∇_κ_∇ϕ(i, j, k, grid, ϕ, closure::ConstantAnisotropicDiffusivity, args...) = (
+      closure.κh * ∂x²_caa(i, j, k, grid, ϕ)
+    + closure.κh * ∂y²_aca(i, j, k, grid, ϕ)
+    + closure.κv * ∂z²_aac(i, j, k, grid, ϕ)
+    )
+=#
 
 # These functions are used to specify Gradient and Value boundary conditions.
 @inline κ₁₁_ccc(i, j, k, grid, closure::ConstantAnisotropicDiffusivity, args...) = closure.κh