@ali-ramadhan 2019-05-27
Cool! Looks pretty neat considering how much tedious math is buried underneath. Thanks for cleaning up the `Model` constructor.

> In the future, we should probably make ConstantIsotropicDiffusivity the default, and remove the option to set the horizontal and vertical diffusion tensor components in the Model constructor.

+1.

Do you want to close #120 once this is merged?