@johncmarshall54 2019-05-27
Shouldn't the default be anisotropic diffusion with the anisotrop scaled by
delx, delz. This will be by far the most common model configuration.

On Mon, May 27, 2019, 4:04 PM Ali Ramadhan <notifications@github.com> wrote:

> *@ali-ramadhan* commented on this pull request.
> ------------------------------
>
> In src/time_steppers.jl
> <https://github.com/climate-machine/Oceananigans.jl/pull/245#discussion_r287873945>
> :
>
> >      @loop for j in (1:grid.Ny; (blockIdx().y - 1) * blockDim().y + threadIdx().y)
>
>          @loop for i in (1:grid.Nx; (blockIdx().x - 1) * blockDim().x + threadIdx().x)
>
> -            apply_z_top_bc!(top_bc, i, j, grid, ϕ, Gϕ, κ, t, iteration, u, v, w, T, S)
>
> -            apply_z_bottom_bc!(bottom_bc, i, j, grid, ϕ, Gϕ, κ, t, iteration, u, v, w, T, S)
>
> +
>
> +               κ_top = κ(i, j, 1,       grid, closure, eos, g, u, v, w, T, S)
>
> +            κ_bottom = κ(i, j, grid.Nz, grid, closure, eos, g, u, v, w, T, S)
>
> +
>
> +               apply_z_top_bc!(top_bc,    i, j, grid, ϕ, Gϕ, κ_top,    t, iteration, u, v, w, T, S)
>
> +            apply_z_bottom_bc!(bottom_bc, i, j, grid, ϕ, Gϕ, κ_bottom, t, iteration, u, v, w, T, S)
>
>
> It looks like if you want to impose a z boundary condition that does not
> depend on κ, you still have to calculate κ using the full closure which
> can be expensive if using an LES closure. Not sure how to get around this,
> probably some clever multiple dispatch?
>
> This is probably fine for now as constant Smagorinsky isn't integrated
> yet, and the performance hit probably isn't big enough to worry about right
> now.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/climate-machine/Oceananigans.jl/pull/245?email_source=notifications&email_token=AKXUEQURKUBCRVSBQUC2LQDPXQ5FPA5CNFSM4HP6AWMKYY3PNVWWK3TUL52HS4DFWFIHK3DMKJSXC5LFON2FEZLWNFSXPKTDN5WW2ZLOORPWSZGOBZZG3JY#pullrequestreview-242380199>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AKXUEQXS3APE4PYPCYWURM3PXQ5FPANCNFSM4HP6AWMA>
> .
>