@glwagner 2019-05-27
@johncmarshall54 we can make that the default. I don't quite understand why that's a common model configuration --- I hope we develop far better closures --- and encourage others to use better closures --- than constant diffusion (even anisotropic... 😏) for all kinds of flow. For my application my default configuration will be anisotropic minimum dissipation (and we can extend this model to accurately mesoscale flows as well).