@glwagner 2019-05-27
> Do you want to close #120 once this is merged?

Let's wait --- I'm working on test cases for `ConstantSmagorinsky` now. The next PR should be the one.