@glwagner 2019-05-28
Another q: should I shower the code in a hailstorm of `@inline`s? @vchuravy, I know we force inline on the GPU, but what about the CPU? Do we force inline the innards of functions annotated with `@launch` as well?