@glwagner 2019-05-28
@ali-ramadhan do you mind running some benchmarks to test for performance regression under this PR?