@glwagner 2019-05-28
Lastly, I am thinking that all the doc strings in `closure_operators.jl` are actually a detriment to readability and understandability. Thoughts?