@ali-ramadhan 2019-05-28
> @ali-ramadhan do you mind running some benchmarks to test for performance regression under this PR?

Looks really bad, still waiting on GPU results but CPU is ~100x slower! We can't merge this as is.

Something must be wrong somewhere. Memory allocations went from `6.00 KiB` per iteration to `1.04 GiB` per iteration. It used to be resolution independent but now the allocation increases with the number of grid points. I'll edit this comment with the full benchmark timings when they're done.

I benchmarked master yesterday so I know it's still performing well (maybe 2-3% slower because of the extra stuff and fixes we recently introduced).