@glwagner 2019-05-28
@ali-ramadhan are those regressions on the CPU or GPU? On the CPU, we might need `inline` (we need inline to elide a number of potential memory allocation points I think). Not sure about GPU.