@ali-ramadhan 2019-05-28
Which version of GPUifyLoops are you using? Maybe post `] status` and we can compare versions. I had issues with v0.2.4 but they showed up in the tests.