@ali-ramadhan 2019-05-28
@glwagner You might be right about the `@inline` thing as the GPU models run reasonably fast, albeit ~30% slower. But not a factor of ~100x.

Before:
```
──────────────────────────────────────────────────────────────────────────────────────────────────
            Oceananigans.jl benchmarks                    Time                   Allocations      
                                                  ──────────────────────   ───────────────────────
                Tot / % measured:                       227s / 45.6%           18.7GiB / 0.06%    

Section                                   ncalls     time   %tot     avg     alloc   %tot      avg
──────────────────────────────────────────────────────────────────────────────────────────────────
256x256x256 static ocean (CPU, Float32)       10    54.4s  52.5%   5.44s   60.0KiB  0.48%  6.00KiB
256x256x256 static ocean (CPU, Float64)       10    36.9s  35.6%   3.69s   77.8KiB  0.62%  7.78KiB
128x128x128 static ocean (CPU, Float32)       10    6.38s  6.16%   638ms   60.0KiB  0.48%  6.00KiB
128x128x128 static ocean (CPU, Float64)       10    4.04s  3.90%   404ms   77.8KiB  0.62%  7.78KiB
 64x 64x 64 static ocean (CPU, Float32)       10    748ms  0.72%  74.8ms   60.0KiB  0.48%  6.00KiB
 64x 64x 64 static ocean (CPU, Float64)       10    412ms  0.40%  41.2ms   77.8KiB  0.62%  7.78KiB
256x256x256 static ocean (GPU, Float64)       10    284ms  0.27%  28.4ms   1.59MiB  12.9%   163KiB
256x256x256 static ocean (GPU, Float32)       10    243ms  0.23%  24.3ms   1.35MiB  11.0%   139KiB
 32x 32x 32 static ocean (CPU, Float32)       10   80.3ms  0.08%  8.03ms   60.0KiB  0.48%  6.00KiB
 32x 32x 32 static ocean (CPU, Float64)       10   45.2ms  0.04%  4.52ms   77.8KiB  0.62%  7.78KiB
128x128x128 static ocean (GPU, Float64)       10   35.9ms  0.03%  3.59ms   1.59MiB  12.9%   163KiB
128x128x128 static ocean (GPU, Float32)       10   32.3ms  0.03%  3.23ms   1.35MiB  11.0%   139KiB
 64x 64x 64 static ocean (GPU, Float64)       10   6.54ms  0.01%   654μs   1.59MiB  12.9%   163KiB
 64x 64x 64 static ocean (GPU, Float32)       10   6.14ms  0.01%   614μs   1.35MiB  11.0%   139KiB
 32x 32x 32 static ocean (GPU, Float64)       10   5.77ms  0.01%   577μs   1.59MiB  12.9%   163KiB
 32x 32x 32 static ocean (GPU, Float32)       10   5.68ms  0.01%   568μs   1.35MiB  11.0%   139KiB
──────────────────────────────────────────────────────────────────────────────────────────────────
```

Now:
```
 ──────────────────────────────────────────────────────────────────────────────────────────────────
                Intermediate output                        Time                   Allocations      
                                                   ──────────────────────   ───────────────────────
                 Tot / % measured:                       116s / 69.4%            115GiB / 81.4%    

 Section                                   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────────────────────────
  64x 64x 64 static ocean (CPU, Float32)       10    70.3s  87.0%   7.03s   83.5GiB  88.9%  8.35GiB
  32x 32x 32 static ocean (CPU, Float32)       10    10.5s  13.0%   1.05s   10.4GiB  11.1%  1.04GiB
 ──────────────────────────────────────────────────────────────────────────────────────────────────
```
```
 ──────────────────────────────────────────────────────────────────────────────────────────────────
             Oceananigans.jl benchmarks                    Time                   Allocations      
                                                   ──────────────────────   ───────────────────────
                 Tot / % measured:                      63.8s / 1.31%           13.6GiB / 0.09%    

 Section                                   ncalls     time   %tot     avg     alloc   %tot      avg
 ──────────────────────────────────────────────────────────────────────────────────────────────────
 256x256x256 static ocean (GPU, Float64)       10    366ms  43.9%  36.6ms   1.79MiB  13.6%   183KiB
 256x256x256 static ocean (GPU, Float32)       10    344ms  41.2%  34.4ms   1.51MiB  11.4%   155KiB
 128x128x128 static ocean (GPU, Float64)       10   49.8ms  5.96%  4.98ms   1.79MiB  13.6%   183KiB
 128x128x128 static ocean (GPU, Float32)       10   48.6ms  5.82%  4.86ms   1.51MiB  11.4%   155KiB
  64x 64x 64 static ocean (GPU, Float64)       10   7.78ms  0.93%   778μs   1.79MiB  13.6%   183KiB
  64x 64x 64 static ocean (GPU, Float32)       10   7.66ms  0.92%   766μs   1.51MiB  11.4%   155KiB
  32x 32x 32 static ocean (GPU, Float32)       10   5.35ms  0.64%   535μs   1.51MiB  11.4%   154KiB
  32x 32x 32 static ocean (GPU, Float64)       10   4.57ms  0.55%   457μs   1.79MiB  13.6%   183KiB
 ──────────────────────────────────────────────────────────────────────────────────────────────────
```